requirements.txt
absl-py==0.7.1
astor==0.8.0
cycler==0.10.0
Cython==0.29.11
decorator==4.4.0
easydict==1.9
editdistance==0.5.3
gast==0.2.2
glog==0.3.1
grpcio==1.21.1
h5py==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lmdb==0.97
networkx==2.3
numpy==1.14.5
opencv-python==4.1.0.25
protobuf==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
python-gflags==3.1.2
pytz==2019.1
scipy==1.3.0
six==1.12.0
tensorboard==1.10.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
tqdm==4.32.1
Werkzeug==0.15.4
wordninja==0.1.5